Update to Completeness/Optimality
The proofs were incomplete ("XXX We should say something..."), and
dealing with it was more difficult that I would have expected.  I added
a lemma that is used in the proofs for completeness (which even then is
more complicated), and optimality (which is now a 1 liner).  I tried to
keep the mathiness down, but there's a limit.

I added a sentence above that is pretty inconspicuous, but y'all be the
judges.